+++ src/components/Status.svelte
@@ -1,8 +1,9 @@
 <script lang="ts">
     import { ship } from "../stores/writeShip";
-    import { evaluate, EvalErrorCode } from "ftlibship";
+    import { evaluate } from "ftlibship";
     import type { IEvaluation } from "ftlibship";
     import { afterUpdate } from "svelte";
+    import { formatEvalError } from "@/lib/statusErrors";
 
     interface ISystem {
         name: string;
@@ -17,55 +18,6 @@
         $ship.cpv = results.cpv;
         delta = $ship.mass - results.mass;
     });
-
-    const errorMsgs: Map<EvalErrorCode, string> = new Map([
-        [EvalErrorCode.NoMass, "You haven't specified a mass for your ship."],
-        [EvalErrorCode.BadMass, "The ship's mass must be between 5 and 300."],
-        [
-            EvalErrorCode.OverMass,
-            "You have equipped more systems than your ship's mass can accommodate.",
-        ],
-        [
-            EvalErrorCode.LowHull,
-            "The ship's hull must be at least 10% of total mass.",
-        ],
-        [
-            EvalErrorCode.OverDCP,
-            "You have allocated too many damage control parties.",
-        ],
-        [
-            EvalErrorCode.OverMarine,
-            "You have allocated too many onboard marines.",
-        ],
-        [
-            EvalErrorCode.OverCrew,
-            "You have overallocated your crew. You may need to add some berths.",
-        ],
-        [
-            EvalErrorCode.OverSpinal,
-            "Your spinal-mounted weapons are too heavy. You can only equip 16 mass of spinal weapons for every 50 total ship mass (rounded up).",
-        ],
-        [
-            EvalErrorCode.OverTurret,
-            "You have too many turrets. You can only have one turret for every 50 total ship mass (rounded up).",
-        ],
-        [
-            EvalErrorCode.DblUID,
-            "At least one of your systems shares a UUID with another. This should not happen in the normal course of using the builder. Did you modify a JSON file?",
-        ],
-        [
-            EvalErrorCode.OverPBL,
-            "You have equipped more plasma bolt launchers than your ship's mass can accommodate.",
-        ],
-        [
-            EvalErrorCode.OverSpinal,
-            "You have equipped more spinal mount weapons than your ship's mass can accommodate.",
-        ],
-        [
-            EvalErrorCode.OverTurret,
-            "You have equipped more turrets than your ship's mass can accommodate.",
-        ],
-    ]);
 
     let modalClearShip: string;
     const clearShip = () => {
@@ -142,7 +94,7 @@
             </div>
             <div class="content">
                 {#each results.errors as e}
-                    <p>{errorMsgs.get(e)}</p>
+                    <p>{formatEvalError(e)}</p>
                 {/each}
             </div>
         {/if}